Added xmlutils for helping parsing fields
This util will reduce the internal code to read the xmlfields,
control the default values, and as a fallback when a new feature
is implemented, and the old files do not contain the search fields.
This should be used to replace all instances of .findall to handle
the errors of it.