Added Theme Editor
New Functionalities!
A theme editor is integrated to BeamerQT!

Initial version functional!

Add code to the preamble corresponding to the theme.
Add files (text files, image files).

During generation of the PDF, the attached files will
be generated in the output directory, and used directly!

Carry with the .xml file meanwhile, add it to the Templates
folder.

Future versions will integrate the template xml content directly
into the BQT file.